+++ NEWS.md
@@ -1,4 +1,4 @@
-# nanonext 0.10.4.9003 (development)
+# nanonext 0.10.4.9004 (development)
 
 #### New Features